Renamed the os2borgerpc_client package to os2borgerpc.client
This commit also makes all of the intra-package imports in
os2borgerpc.client relative
