+++ os2borgerpc_client/bin/get_os2borgerpc_config
@@ -2,7 +2,7 @@
 
 import sys
 
-from os2borgerpc_client.config import get_config
+from os2borgerpc.client.config import get_config
 
 DEBUG = True  # TODO: Get from command line option
