+++ os2borgerpc_client/bin/jobmanager
@@ -3,7 +3,7 @@
 import os
 import sys
 
-from os2borgerpc_client.jobmanager import update_and_run
+from os2borgerpc.client.jobmanager import update_and_run
 
 if not os.geteuid() == 0:
     sys.exit("\n Only root can run this program. \n")
