+++ os2borgerpc_client/bin/os2borgerpc_push_config_keys
@@ -2,8 +2,8 @@
 
 import sys
 
-from os2borgerpc_client.config import OS2borgerPCConfig
-from os2borgerpc_client.admin_client import get_default_admin
+from os2borgerpc.client.config import OS2borgerPCConfig
+from os2borgerpc.client.admin_client import get_default_admin
 try:
     config = OS2borgerPCConfig()
     to_push = {}
