+++ os2borgerpc_client/bin/os2borgerpc_register_in_admin
@@ -3,8 +3,8 @@
 import sys
 import urllib.parse
 
-from os2borgerpc_client.config import OS2borgerPCConfig
-from os2borgerpc_client.admin_client import OS2borgerPCAdmin
+from os2borgerpc.client.config import OS2borgerPCConfig
+from os2borgerpc.client.admin_client import OS2borgerPCAdmin
 
 
 config = OS2borgerPCConfig()
