+++ os2borgerpc_client/bin/os2borgerpc_upload_dist_packages
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from os2borgerpc_client.utils import upload_dist_packages
+from os2borgerpc.client.utils import upload_dist_packages
 
 
 if __name__ == '__main__':
