+++ os2borgerpc_client/bin/os2borgerpc_upload_packages
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from os2borgerpc_client.utils import upload_packages
+from os2borgerpc.client.utils import upload_packages
 
 if __name__ == '__main__':
     upload_packages()
